[cmake] optional demos + reorganize demos compilation in subdirectory demos
